chore: add .pre-commit-config.yaml file docs(README.md): add badge for pre-commit & fix style of other badges docs(README.md): fix incorrect CONTRIBUTING linkinkg branch name
The .pre-commit-config.yaml file is added to the project, which includes
hooks for checking YAML and JSON files. This will help ensure the quality
and consistency of these file formats. Additionally, the README.md file is
updated to fix the link to the CONTRIBUTING guide, directing users to the
CONTRIBUTING.md file in the main branch. This ensures that contributors
have access to the most up-to-date instructions for helping the project
grow.